Code refactor
Moved code out of policy_iteration2 for reuse in policy_iteration
