+++ Forward_Backward_Algiorithm_wikipedia.py
@@ -1,10 +1,59 @@
-#Original Source code from wikipedia
+#SOURCE: https://en.wikipedia.org/wiki/Forward%E2%80%93backward_algorithm
 import numpy as np
 import pandas as pd
 import math
 
+def fwd_bkw_custom(observations, states, start_prob, trans_prob, emm_prob, end_st):
+    # Forward part of the algorithm
+    fwd = []
+    for i, observation_i in enumerate(observations):
+        f_curr = {}
+        for st in states:
+            if i == 0:
+                # base case for the forward part
+                prev_f_sum = start_prob[st]
+            else:
+                prev_f_sum = sum(f_prev[k] * trans_prob[k][st] for k in states)
+
+            f_curr[st] = emm_prob[st][observation_i] * prev_f_sum
+
+        fwd.append(f_curr)
+        f_prev = f_curr
+
+    p_fwd = sum(f_curr[k] * trans_prob[k][end_st] for k in states)
+
+    # Backward part of the algorithm
+    bkw = []
+    for i, observation_i_plus in enumerate(reversed(observations[1:] + [None,])):
+        b_curr = {}
+        for st in states:
+            if i == 0:
+                # base case for backward part
+                b_curr[st] = trans_prob[st][end_st]
+            else:
+                b_curr[st] = sum(trans_prob[st][l] * emm_prob[l][observation_i_plus] * b_prev[l] for l in states)
+        bkw.insert(0,b_curr)
+        b_prev = b_curr
+
+    p_bkw = sum(start_prob[l] * emm_prob[l][observations[0]] * b_curr[l] for l in states)
+
+    # Merging the two parts
+    posterior = []
+    for i in range(len(observations)):
+    	#this was added because the state transition distribution had rows with all 0s for the terminal state.
+    	#you have to set 1.0 to the terminal state when transitioning from the terminal state
+    	#if not p_fwd:
+    	#	posterior.append({st: 0.0 for st in states})
+    	#else:
+    	#	posterior.append({st: fwd[i][st] * bkw[i][st] / p_fwd for st in states})
+    	#if the terminal state cycles to itself then you don't need to make the check above.
+    	posterior.append({st: fwd[i][st] * bkw[i][st] / p_fwd for st in states})
+    print(p_fwd)
+    print(p_bkw)
+    #https://davidamos.dev/the-right-way-to-compare-floats-in-python/
+    assert math.isclose(p_fwd, p_bkw)     
+    return fwd, bkw, posterior
 def fwd_bkw(observations, states, start_prob, trans_prob, emm_prob, end_st):
-    """Forward–backward algorithm."""
     # Forward part of the algorithm
     fwd = []
     for i, observation_i in enumerate(observations):
@@ -47,6 +96,7 @@ def fwd_bkw(observations, states, start_prob, trans_prob, emm_prob, end_st):
     	#	posterior.append({st: 0.0 for st in states})
     	#else:
     	#	posterior.append({st: fwd[i][st] * bkw[i][st] / p_fwd for st in states})
+    	#if the terminal state cycles to itself then you don't need to make the check above.
     	posterior.append({st: fwd[i][st] * bkw[i][st] / p_fwd for st in states})
     print(p_fwd)
     print(p_bkw)
