+++ helpers.py
@@ -1,0 +1,174 @@
+import numpy as np
+import math
+import random
+
+CMP_DELTA = 0.000001
+
+def information_gain(Q, P, s, path_probability):
+    """Calculate a bound for expected(?) information gain given one execution of viturbi on one string of observations
+    Q is the "prior" distribution over all states for a certain time
+    P is the "posterior" distribution over all states for a certain time
+    s is the state we are interested in
+    path_probability is the probability of a path (pr of observation string * pr of most likely path given by viterbi)"""
+
+    # calculate information gain (relative entropy, or Kullback–Leibler divergence)
+    # D_{kl}(P || Q) = sum over all outcomes of Pr(outcome)*log2(P(outcome)/Q(outcome))
+
+    rel_entropy = 0.0
+    # If Q[s] == 0 or Q[s] == 1, then there's no uncertainty
+    if Q[s] < CMP_DELTA or Q[s] > 1.0 - CMP_DELTA:
+        return (0.0, 0.0)
+    # If P[s] = 0, then math works out ok, since 0log0 == 0
+    elif P[s] < CMP_DELTA:
+        rel_entropy = (1.0 - P[s]) * math.log2((1.0 - P[s]) / (1.0 - Q[s]))
+    # otherwise, use standard formula
+    else:
+        rel_entropy = P[s] * math.log2(P[s] / Q[s]) + \
+            (1.0 - P[s]) * math.log2((1.0 - P[s]) / (1.0 - Q[s]))
+
+    max_remaining_info = 0.0
+    if Q[s] > 0.5:
+        # max info is when P=0.0
+        max_remaining_info = (1.0 - 0.0) * math.log2(1.0 / Q[s])
+    else:
+        # max info is when P=1.0
+        max_remaining_info = 1.0 * math.log2(1.0 / Q[s])
+
+    # what we know based on paths we tested.
+    known_rel_entropy = rel_entropy
+    # worst case expected entropy, given what we know.
+    worst_expected_entropy = path_probability * rel_entropy + (1.0 - path_probability) * max_remaining_info
+    return (known_rel_entropy, worst_expected_entropy)
+
+def action_to_str(a):
+    if a == -1:
+        return "DONE"
+    elif a == 0:
+        return "^"
+    elif a == 1:
+        return "<"
+    elif a == 2:
+        return "v"
+    elif a == 3:
+        return ">"
+    return "#"
+
+def action_to_str2(a):
+    result = '#'
+    if a == -1:
+        result = "DONE"
+    elif a == 0:
+        result = ">"
+    elif a == 1:
+        result = "|"
+    return result
+
+def take_action(curr_state, action, T):
+    """Return the next state and the given current state and the action chosen
+
+    """
+    coin  = random.random()
+    # coin = 0.5
+    # 12 possible next states
+    next_states = T[curr_state, : , int(action)]
+    prob_counter = 0.0
+    # randomly take next action based on weights
+    for state, prob in enumerate(next_states):
+        if coin < prob_counter + prob:
+            return state
+        prob_counter += prob
+    return -1
+
+def execute_policy(p, T, start, max_t):
+    """Place an agent in the environment and generate a stream of actions
+
+    """
+    curr_state = start
+    output = []
+    # no longer than max_t steps
+    for i in range(max_t):
+        output.append(int(p[curr_state]))
+        if p[curr_state] == -1:
+            break
+        curr_state = take_action(curr_state, p[curr_state], T)
+    return output
+
+def to_markov_chain(p, T, max_t):
+    result = [[0]*max_t]*max_t
+    for t in range(max_t):
+        if not np.isnan(p[t]):
+            result[t] = [row[p[t]] for row in T[t][:]]
+    return result
+
+def get_expected_visits(states, start_p, T, p, t):
+    """Get number of extpected visits of each state after t steps
+    with no information about observations
+
+    states : state indices
+    start_p : initial probability distribution
+    T : original transition matrix
+    p : policy
+    t : time interval
+    """
+    # build new transition matrix from policy
+    trans_p = []
+    for i in range(12):
+        trans_p.append([0.0 for j in range(12)])
+        if not np.isnan(p[i]) and not p[i] == -1:
+            for j in range(12):
+                trans_p[i][j] = T[i, j, int(p[i])]
+        elif p[i] == -1:
+            # if at a terminal, then consider that you are at this state for all remaining time
+            trans_p[i][i] = 1.0
+
+    # initial distribution tells us where we will be at time=0
+    curr_p = [start_p[j] for j in range(12)]
+    print("time=%d : %s" % (0, ', '.join(["%.2f" % curr_p[st] for st in states]) + ": sum=%.2f" % sum(curr_p)))
+    for i in range(1,t+1):
+        next_p = [0.0 for j in range(12)]
+        for st in states:
+            for next_st in states:
+                next_p[next_st] += curr_p[st] * trans_p[st][next_st]
+        for st in states:
+            curr_p[st] = next_p[st]
+        print("time=%d : %s" % (i, ', '.join(["%.2f" % curr_p[st] for st in states]) + ": sum=%.2f" % sum(curr_p)))
+    return curr_p
+
+def dijkstra(trans_p, start, goal):
+    distances = [math.inf for i in range(12)]
+    previous = [math.nan for i in range(12)]
+    start = 4
+    min_heap = [(0,start)]
+    distances[start] = 0
+    previous[start] = -1
+    goal = 2
+    max_time = 11
+    while len(min_heap) != 0:
+        value, curr_state = heapq.heappop(min_heap)
+        if curr_state == goal:
+            break
+        for i in range(12):
+            prob = trans_p[curr_state][i]
+            if prob != 0:
+                distance = - math.log(prob)
+                alt = distances [curr_state] + distance
+                if alt < distances[i]:
+                    distances[i] = alt
+                    previous[i] = curr_state
+                    heapq.heappush(min_heap,(alt, i))
+
+    return previous
+
+def generate_naive_paths():
+    """Generate all action sequences, then narrow down
+    based on 
+
+    Goal: gain intuition on how to create a smarter generator of observation strings
+    """
+    pass
+
+def main():
+    print("Calling main function in helpers.py")
+
+if __name__ == "__main__":
+    main()
