+++ policy_iteration.py
@@ -24,23 +24,16 @@
 #Example of the policy iteration algorithm.
 
 import numpy as np
+import pandas as pd
 import random
 import math
+import matplotlib.pyplot as plt
+from scipy.stats import entropy
+from scipy.special import rel_entr, kl_div
 
-CMP_DELTA = 0.000001
-
-def action_to_str(a):
-    if a == -1:
-        return "DONE"
-    elif a == 0:
-        return "^"
-    elif a == 1:
-        return "<"
-    elif a == 2:
-        return "v"
-    elif a == 3:
-        return ">"
-    return "#"
+import Forward_Backward_Algiorithm_wikipedia as fb
+import Viterbi_Algorithm_wikipedia as vt
+import helpers as hlp
 
 def return_policy_evaluation(p, u, r, T, gamma):
     for s in range(12):
@@ -95,48 +88,22 @@ def print_policy(p, shape):
         policy_string += '\n'
     print(policy_string)
 
-def take_action(curr_state, action, T):
-    """Return the next state and the given current state and the action chosen
-
-    """
-    coin  = random.random()
-    # coin = 0.5
-    # 12 possible next states
-    next_states = T[curr_state, : , int(action)]
-    prob_counter = 0.0
-    # randomly take next action based on weights
-    for state, prob in enumerate(next_states):
-        if coin < prob_counter + prob:
-            return state
-        prob_counter += prob
-    return -1
-
-def execute_policy(p, T, start, max_t):
-    """Place an agent in the environment and generate a stream of actions
-
-    """
-    curr_state = start
-    output = []
-    # no longer than max_t steps
-    for i in range(max_t):
-        output.append(int(p[curr_state]))
-        if p[curr_state] == -1:
-            break
-        curr_state = take_action(curr_state, p[curr_state], T)
-    return output
-
-def main_iterative():
+def main_iterative(obs = []):
     """Finding the solution using the iterative approach
 
     """
     gamma = 0.999
     iteration = 0
+    print("================= LOADING TRANSITIONAL MATRIX ==================")
     T = np.load("T.npy")
-
+    print(T)
+    print("====================== POLICY ITERATION ========================")
     #Generate the first policy randomly
     # Nan=Nothing, -1=Terminal, 0=Up, 1=Left, 2=Down, 3=Right
     p = np.random.randint(0, 4, size=(12)).astype(np.float32)
+    #Obstacles
     p[5] = np.NaN
+    #Terminal States
     p[3] = p[7] = -1
 
     #Utility vectors
@@ -167,19 +134,49 @@ def main_iterative():
                 if a != p[s]: p[s] = a
         print_policy(p, shape=(3,4))
 
-    print("=================== FINAL RESULT ==================")
+    print("================ POLICY ITERATION FINAL RESULT =================")
     print("Iterations: " + str(iteration))
     print("Delta: " + str(delta))
     print("Gamma: " + str(gamma))
     print("Epsilon: " + str(epsilon))
-    print("===================================================")
+    print("================================================================")
     print(u[0:4])
     print(u[4:8])
     print(u[8:12])
-    print("===================================================")
+    print("================================================================")
     print_policy(p, shape=(3,4))
-    print("===================================================")
-    start_pos = 11
+    print("===================== A Priori Analysis ======================")
+    print("==================MDP + Policy = Markov Chain ==================")
+    print("Policy: ")
+    policy = [np.NaN if np.isnan(i) else int(i) for i in p]
+    print("Markov Chain:")
+    markov_chain = hlp.to_markov_chain(policy, T, 12)
+    markov_chain_df = pd.DataFrame(markov_chain)
+    print(markov_chain_df)
+
+    #set terminal state
+    markov_chain[7][3]=1.0
+    markov_chain[7][7]=1.0
+    #set start state
+    state = [
+        [0.0,0.0,0.0,0.0,
+         1.0,0.0,0.0,0.0,
+         0.0,0.0,0.0,0.0]]
+    state_history = [state[0]]
+    for x in range(20):
+        next_state = [[sum(a * b for a, b in zip(state_row, markov_chain_col))
+                        for markov_chain_col in zip(*markov_chain)]
+                                for state_row in state]
+        state_history.append(next_state[0])
+        state = next_state
+    print("Stationary Distribution")
+    print(state)
+    state_history_df = pd.DataFrame(state_history)
+    #state_history_df.plot()
+    #plt.show()
+    print(state_history_df)
+    print("================================================================")
+    start_pos = 8
     states = [i for i in range(12)]
     start_p = [0.0 for i in range(12)]
     start_p[start_pos] = 1.0
@@ -199,196 +196,94 @@ def main_iterative():
         # TODO - make nondeterministic policy possible
         if not np.isnan(p[i]):
             emit_p[i][int(p[i])+1] = 1.0
-    
-    print("====================== A Priori Analysis ====================")
+    print("========================= Dijkstra's =========================")
+    # print("Distances")
+    # print(distances)
+    # print(trans_p)
     interesting_time = 4
     interesting_state = 3
-    prior_expected_visits = get_expected_visits(states, start_p, T, p, interesting_time)
+    prior_expected_visits = hlp.get_expected_visits(states, start_p, T, p, interesting_time)
     print("Expected visits: \n" + ', '.join(["%.2f" % prior_expected_visits[st] for st in states]))
     print("Sum of expected visits should = 1 + t. %.2f == %d." % (sum(prior_expected_visits), 1+interesting_time) )
-    print("=================== EXEC  POLICY ==================")
-    obs = execute_policy(p, T, start_pos, 12)
+    if not obs:
+        print("====================== Executing Policy ======================")
+        obs = hlp.execute_policy(p, T, start_pos, 12)
+        #obs = [0,0,1,-1]
+    else:
+        print("===================== Not Executing Policy ===================")
+    print(obs)
     s = "["
     for a in obs:
-        s += action_to_str(a) + ", "
+        s += hlp.action_to_str(a) + ", "
     if len(s) > 1:
         print(s[:-2] + "]")
     else:
         print("[]")
     # obs needs positive indices for viterbi alg implementation below
+    #obs_original = obs
     obs = [obs[i]+1 for i in range(len(obs))]
-    print("====================== VITERBI ====================")
-    (dp_table, max_path_prob) = viterbi(obs, states, start_p, trans_p, emit_p)
+    
+    end_state = 7
+    trans_p[end_state][end_state]= 1.0 #setting terminal state?
+    
+    trans_p_df = pd.DataFrame(trans_p)
+    emit_p_df = pd.DataFrame(emit_p)
+    print(obs)
+    print(states)
+    print(start_p)
+    print(trans_p_df)
+    print(emit_p_df)
+    print("=========================== VITERBI ==========================")
+    (dp_table, max_path_prob) = vt.viterbi_custom(obs, states, start_p, trans_p, emit_p)
     if interesting_time >= len(dp_table):
         print("Actual execution did not go as long as %d steps. How to handle information gain here?" % interesting_time)
     else:
         post_expected_visits = [dp_table[interesting_time][st]["prob"] for st in states]
         print("Actual expected visits given single execution: \n" + ', '.join(["%.2f" % post_expected_visits[st] for st in states]))
         print("====================== INFORMATION GAIN ====================")
-        ig = information_gain(prior_expected_visits, post_expected_visits, interesting_state, max_path_prob)
-        print("Information Gain on state=%d and time=%d: %.2f" % (interesting_state, interesting_time, ig))
-
-def information_gain(Q, P, s, path_probability):
-    """Calculate a bound for expected(?) information gain given one execution of viturbi on one string of observations
-    Q is the "prior" distribution over all states for a certain time
-    P is the "posterior" distribution over all states for a certain time
-    s is the state we are interested in
-    path_probability is the probability of a path (pr of observation string * pr of most likely path given by viterbi)"""
-
-    # calculate information gain (relative entropy, or Kullback–Leibler divergence)
-    # D_{kl}(P || Q) = sum over all outcomes of Pr(outcome)*log2(P(outcome)/Q(outcome))
-
-    rel_entropy = 0.0
-    # If Q[s] == 0 or Q[s] == 1, then there's no uncertainty
-    if Q[s] < CMP_DELTA or Q[s] > 1.0 - CMP_DELTA:
-        return (0.0, 0.0)
-    # If P[s] = 0, then math works out ok, since 0log0 == 0
-    elif P[s] < CMP_DELTA:
-        rel_entropy = (1.0 - P[s]) * math.log2((1.0 - P[s]) / (1.0 - Q[s]))
-    # otherwise, use standard formula
-    else:
-        rel_entropy = P[s] * math.log2(P[s] / Q[s]) + \
-            (1.0 - P[s]) * math.log2((1.0 - P[s]) / (1.0 - Q[s]))
-
-    max_remaining_info = 0.0
-    if Q[s] > 0.5:
-        # max info is when P=0.0
-        max_remaining_info = (1.0 - 0.0) * math.log2(1.0 / Q[s])
-    else:
-        # max info is when P=1.0
-        max_remaining_info = 1.0 * math.log2(1.0 / Q[s])
+        #ig = hlp.information_gain(prior_expected_visits, post_expected_visits, interesting_state, max_path_prob)
+        #print("Information Gain on state=%d and time=%d: %.2f" % (interesting_state, interesting_time, ig))
+    print("=========================== Forward Backward ==========================")
+    result = fb.fwd_bkw_custom(obs, states, start_p, trans_p, emit_p, end_state)
+    for line in result:
+        print(*line)
+    print('##FORWARD##')
+    print(pd.DataFrame(result[0]))
+    print('##BACKWARD##')
+    print(pd.DataFrame(result[1]))
+    print('##POSTERIOR##')
+    print(pd.DataFrame(result[2]))
     
-    # what we know based on paths we tested.
-    known_rel_entropy = rel_entropy
-    # worst case expected entropy, given what we know.
-    worst_expected_entropy = path_probability * rel_entropy + (1.0 - path_probability) * max_remaining_info
-    return (known_rel_entropy, worst_expected_entropy)
-
-def get_expected_visits(states, start_p, T, p, t):
-    """Get number of extpected visits of each state after t steps
-    with no information about observations
-
-    states : state indices
-    start_p : initial probability distribution
-    T : original transition matrix
-    p : policy
-    t : time interval
-    """
-    # build new transition matrix from policy
-    trans_p = []
-    for i in range(12):
-        trans_p.append([0.0 for j in range(12)])
-        if not np.isnan(p[i]) and not p[i] == -1:
-            for j in range(12):
-                trans_p[i][j] = T[i, j, int(p[i])]
-        elif p[i] == -1:
-            # if at a terminal, then consider that you are at this state for all remaining time
-            trans_p[i][i] = 1.0
+    #p = [i.values() for i in result[2]]
+    #convert dictionary to list
+    p = []
+    for i in range(len(obs)):
+    	val_ls = []
+    	for key, val in result[2][i].items():
+    		val_ls.append(val)
+    	p.append(val_ls)
+    q = state_history[:len(obs)]
     
-    # initial distribution tells us where we will be at time=0
-    curr_p = [start_p[j] for j in range(12)]
-    print("time=%d : %s" % (0, ', '.join(["%.2f" % curr_p[st] for st in states]) + ": sum=%.2f" % sum(curr_p)))
-    for i in range(1,t+1):
-        next_p = [0.0 for j in range(12)]
-        for st in states:
-            for next_st in states:
-                next_p[next_st] += curr_p[st] * trans_p[st][next_st]
-        for st in states:
-            curr_p[st] = next_p[st]
-        print("time=%d : %s" % (i, ', '.join(["%.2f" % curr_p[st] for st in states]) + ": sum=%.2f" % sum(curr_p)))
-    return curr_p
-
-def generate_naive_paths():
-    """Generate all action sequences, then narrow down
-    based on 
-
-    Goal: gain intuition on how to create a smarter generator of observation strings
-    """
-    pass
-
-# SOURCE: https://en.wikipedia.org/wiki/Viterbi_algorithm
-def viterbi(obs, states, start_p, trans_p, emit_p):
-    V = [{}]
-    # TODO - modification? For us, first observation is no different than the second
-    for st in states:
-        V[0] [st] = {"prob": start_p[st] * emit_p[st] [obs[0]], "prev": None}
-    # Run Viterbi when t > 0
-    for t in range(1, len(obs)):
-        V.append({})
-        prob_sum = 0.0
-        for st in states:
-            max_tr_prob = V[t - 1] [states[0]] ["prob"] * trans_p[states[0]] [st]
-            # print("max prob = ", max_tr_prob)
-            prev_st_selected = states[0]
-            for prev_st in states[1:]:
-                tr_prob = V[t - 1] [prev_st] ["prob"] * trans_p[prev_st] [st]
-                # print("testing against", tr_prob)
-                if tr_prob > max_tr_prob:
-                    max_tr_prob = tr_prob
-                    prev_st_selected = prev_st
-
-            max_prob = max_tr_prob * emit_p[st][obs[t]]
-            V[t] [st] = {"prob": max_prob, "prev": prev_st_selected}
-            prob_sum += max_prob
-        
-        # Update probabilities to sum to 1.0 at each time
-        for st in states:
-            V[t] [st] ["prob"] /= prob_sum
-
-    # Back-propagate the answers
-    # ASSUMES YOU KNOW YOU ENDED AT A GOAL STATE
-    # Is this like running a version of viturbi backwards now?
-    # 1. re-weight previous probabilities based on new info
-    #       Easy way - when pr=1.0 at time=t, eliminate non-adjacent states at t-1 and so on
-    #       What if not pr=1.0 at time=t? TODO - think about this.
-    # 2. update to sum to 1.0
-    for t in range(len(obs) - 1, 0, -1):
-        for st in states:
-            if V[t] [st] ["prob"] > 0.999999 and V[t] [st] ["prob"] < 1.000001:
-                prob_sum = 0.0
-                for prev_st in states:
-                    # all non-adjacent states probability = 0
-                    if trans_p[prev_st] [st] < 0.000001:
-                        V[t-1] [prev_st] ["prob"] = 0.0
-                    prob_sum += V[t-1] [prev_st] ["prob"]
-                # Update probabilities to sum to 1.0
-                for prev_st in states:
-                    V[t-1] [prev_st] ["prob"] /= prob_sum
-
-    for line in dptable(V):
-        print(line)
-
-    opt = []
-    max_prob = 0.0
-    best_st = None
-    # Get most probable state and its backtrack
-    for st, data in V[-1].items():
-        if data["prob"] > max_prob:
-            max_prob = data["prob"]
-            best_st = st
-    opt.append(best_st)
-    previous = best_st
-
-    path_prob = max_prob
-    # Follow the backtrack till the first observation
-    for t in range(len(V) - 2, -1, -1):
-        opt.insert(0, V[t + 1] [previous] ["prev"])
-        path_prob *= V[t + 1] [previous] ["prob"]
-        previous = V[t + 1] [previous] ["prev"]
-
-    print ("The steps of states are ", opt, " with highest probability of ", path_prob)
-    return (V, path_prob)
-
-def dptable(V):
-    # Print a table of steps from dictionary
-    yield " " * 5 + "     ".join(("%3d" % i) for i in range(len(V)))
-    for state in V[0]:
-        yield "%.7s: " % state + " ".join("%.7s" % ("%lf" % v[state] ["prob"]) for v in V)
+    #https://stackoverflow.com/questions/63369974/3-functions-for-computing-relative-entropy-in-scipy-whats-the-difference
+    print('##ACTUAL DISTRIBUTION##')
+    print(pd.DataFrame(p))
+    print('##REFERENCE DISTRIBUTION##')
+    print(pd.DataFrame(q))
+    print('##ENTROPY##')
+    for i in range(len(p)):
+        print(entropy(p[i], q[i]))
+    print('##RELATIVE ENTROPY##')
+    for i in range(len(p)):
+        print(rel_entr(p[i], q[i]))
+        print(sum(rel_entr(p[i], q[i])))
+    print('##KL DIVERGENCE##')
+    for i in range(len(p)):
+        print(kl_div(p[i], q[i]))
+        print(sum(kl_div(p[i], q[i])))
 
 def main():
     main_iterative()
     #main_linalg()
-
 
 if __name__ == "__main__":
     main()
